Adding Auth module

+++ app/api/routes/listSongs.js
@@ -3,29 +3,55 @@ const Repository = require('../repositories/ListRepository');
 const Service = require('../services/ListService');
 const Validator = require('../services/Validator');
 const  Validations = require('../services/Validation');
-
+const UserService = require('./../services/UserService');
 
 const router = express.Router();
-const listSongService = new  Service(new Repository(), new Validator([new Validations.ValidationTitle(), new Validations.ValidationAlbum(), new Validations.ValidationArtist()]), new Validator([ new Validations.ValidationList()]));
+//const jwt = require('jsonwebtoken');
 
-router.get('/' ,(request, response) => {
-    response.send(listSongService.GetLists(request.query));
+const userService = new UserService.userService('secretkey');
+const listSongService = new  Service( new Repository(), new Validator([new Validations.ValidationTitle(), new Validations.ValidationAlbum(), new Validations.ValidationArtist()]), new Validator([ new Validations.ValidationList()]));
+
+router.get('/' , UserService.verify ,(request, response) => {
+
+    if(userService.Verify(request.token) !== undefined){
+        response.send(listSongService.GetLists(request.query));
+    }else{
+        response.sendStatus(403);
+    }
 });
 
 router.post('/', (request, response) => {
-    response.send(listSongService.AddList(request.body['name']));
+
+    if(userService.Verify(request.token) !== undefined){
+        response.send(listSongService.AddList(request.body['name']));
+    }else{
+        response.sendStatus(403);
+    }
 });
 
 router.put('/', (request, response) => {
-    response.send(listSongService.AddSong(request.body['name'], request.body['newSong']));
+
+    if(userService.Verify(request.token) !== undefined){
+        response.send(listSongService.AddSong(request.body['name'], request.body['newSong']));
+    }else{
+        response.sendStatus(403);
+    }
 });
 
 router.delete('/', (request, response) => {
-    response.send(listSongService.DeleteList(request.body['name']));
+    if(userService.Verify(request.token) !== undefined){
+        response.send(listSongService.DeleteList(request.body['name']));
+    }else{
+        response.sendStatus(403);
+    }
 });
 
 router.delete('/song', (request, response) => {
-    response.send(listSongService.DeleteSong(request.body['name'], request.body['song']))
+    if(userService.Verify(request.token) !== undefined){
+        response.send(listSongService.DeleteSong(request.body['name'], request.body['song']))
+    }else{
+        response.sendStatus(403);
+    }
 });
